# use variant keys for particular selector