Currently, ALL variant keys are ALWAYS passed to selectKey().

We only want to pass variant (unique) keys for a particular selector.

```console
$ benchstat a.txt b.txt                       
goos: darwin
goarch: arm64
pkg: go.expect.digital/mf2/template
cpu: Apple M2 Max
                   │    a.txt    │               b.txt                │
                   │   sec/op    │   sec/op     vs base               │
Template_Sprint-12   2.428µ ± 0%   2.333µ ± 0%  -3.91% (p=0.000 n=10)

                   │    a.txt     │                b.txt                │
                   │     B/op     │     B/op      vs base               │
Template_Sprint-12   4.297Ki ± 0%   4.141Ki ± 0%  -3.64% (p=0.000 n=10)

                   │   a.txt    │               b.txt               │
                   │ allocs/op  │ allocs/op   vs base               │
Template_Sprint-12   86.00 ± 0%   82.00 ± 0%  -4.65% (p=0.000 n=10)
```